Extract remaining shared calculator functionality
