+++ next/components/forms/segments/PropertyTaxCalculator/PropertyTaxCalculator.tsx
@@ -3,9 +3,11 @@ import {
   CustomComponentPropertyCalculator,
   CustomComponentPropertyCalculatorProps,
 } from '@schema-generator/generator/uiOptionsTypes'
-import { getTaxCalculatorExpression } from '@schema-generator/tax-form/calculators'
+import {
+  calculateTaxCalculatorExpression,
+  getTaxCalculatorExpression,
+} from '@schema-generator/tax-form/calculators'
 import cx from 'classnames'
-import clone from 'lodash/clone'
 import get from 'lodash/get'
 import React, { useMemo } from 'react'
 import { useNumberFormatter } from 'react-aria'
@@ -59,25 +61,13 @@ const Calculator = ({
   const expression = useMemo(() => getTaxCalculatorExpression(formula), [formula])
 
   const value = useMemo(() => {
-    try {
-      const path = getPath(widget?.id ?? '', dataContextLevelsUp)
-      const dataAtPath = path ? (get(formData, path) as GenericObjectType) : null
-      if (dataAtPath == null) {
-        return null
-      }
-
-      // It is not in the documentation, but `expr-eval` mutates the original data!
-      const clonedData = clone(dataAtPath)
-      const evaluated = expression?.evaluate(clonedData)
-
-      if (!Number.isFinite(evaluated)) {
-        return null
-      }
-
-      return evaluated as number
-    } catch (error) {
+    const path = getPath(widget?.id ?? '', dataContextLevelsUp)
+    const dataAtPath = path ? (get(formData, path) as GenericObjectType) : null
+    if (dataAtPath == null) {
       return null
     }
+
+    return calculateTaxCalculatorExpression(expression, dataAtPath)
   }, [expression, formData, dataContextLevelsUp, widget?.id])
 
   const wrapperClassName = cx('inline-flex items-center justify-start gap-8 self-stretch py-5', {
